+++ src/mysql/statement_base.h
@@ -37,7 +37,7 @@ protected:
 public:
 
    virtual void       reset( void ) override;
-   virtual uint32_t   execute( void ) override;
+   virtual uint64_t   execute( void ) override;
    virtual db::result result( void ) override;
 
 };
